Speed up Beam presubmits

+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/spanner/SpannerWriteIT.java
@@ -28,6 +28,7 @@ import com.google.cloud.spanner.Database;
 import com.google.cloud.spanner.DatabaseAdminClient;
 import com.google.cloud.spanner.DatabaseId;
 import com.google.cloud.spanner.Dialect;
+import com.google.cloud.spanner.KeySet;
 import com.google.cloud.spanner.Mutation;
 import com.google.cloud.spanner.ResultSet;
 import com.google.cloud.spanner.SessionPoolOptions;
@@ -61,8 +62,9 @@ import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.base.Predicate
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.base.Throwables;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.hamcrest.TypeSafeMatcher;
-import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -105,15 +107,15 @@ public class SpannerWriteIT {
     void setTable(String value);
   }
 
-  private Spanner spanner;
-  private DatabaseAdminClient databaseAdminClient;
-  private SpannerTestPipelineOptions options;
-  private String databaseName;
-  private String pgDatabaseName;
-  private String project;
+  private static Spanner spanner;
+  private static DatabaseAdminClient databaseAdminClient;
+  private static SpannerTestPipelineOptions options;
+  private static String databaseName;
+  private static String pgDatabaseName;
+  private static String project;
 
-  @Before
-  public void setUp() throws Exception {
+  @BeforeClass
+  public static void setUpTestEnvironment() throws Exception {
     PipelineOptionsFactory.register(SpannerTestPipelineOptions.class);
     options = TestPipeline.testingPipelineOptions().as(SpannerTestPipelineOptions.class);
 
@@ -152,15 +154,15 @@ public class SpannerWriteIT {
                     + "  Key           INT64,"
                     + "  Value         STRING(MAX) NOT NULL,"
                     + ") PRIMARY KEY (Key)"));
-    op.get();
-    databaseAdminClient
-        .createDatabase(
+    OperationFuture<Database, CreateDatabaseMetadata> pgOp =
+        databaseAdminClient.createDatabase(
             databaseAdminClient
                 .newDatabaseBuilder(DatabaseId.of(project, options.getInstanceId(), pgDatabaseName))
                 .setDialect(Dialect.POSTGRESQL)
                 .build(),
-            Collections.emptyList())
-        .get();
+            Collections.emptyList());
+    op.get();
+    pgOp.get();
     databaseAdminClient
         .updateDatabaseDdl(
             options.getInstanceId(),
@@ -176,7 +178,18 @@ public class SpannerWriteIT {
         .get();
   }
 
-  private String generateDatabaseName() {
+  @Before
+  public void setUp() {
+    Mutation deleteAll = Mutation.delete(options.getTable(), KeySet.all());
+    spanner
+        .getDatabaseClient(DatabaseId.of(project, options.getInstanceId(), databaseName))
+        .writeAtLeastOnce(Collections.singletonList(deleteAll));
+    spanner
+        .getDatabaseClient(DatabaseId.of(project, options.getInstanceId(), pgDatabaseName))
+        .writeAtLeastOnce(Collections.singletonList(deleteAll));
+  }
+
+  private static String generateDatabaseName() {
     String random =
         RandomUtils.randomAlphaNumeric(
             MAX_DB_NAME_LENGTH - 4 - options.getDatabaseIdPrefix().length());
@@ -390,11 +403,15 @@ public class SpannerWriteIT {
     result.waitUntilFinish();
   }
 
-  @After
-  public void tearDown() throws Exception {
-    databaseAdminClient.dropDatabase(options.getInstanceId(), databaseName);
-    databaseAdminClient.dropDatabase(options.getInstanceId(), pgDatabaseName);
-    spanner.close();
+  @AfterClass
+  public static void tearDown() throws Exception {
+    if (databaseAdminClient != null) {
+      databaseAdminClient.dropDatabase(options.getInstanceId(), databaseName);
+      databaseAdminClient.dropDatabase(options.getInstanceId(), pgDatabaseName);
+    }
+    if (spanner != null) {
+      spanner.close();
+    }
   }
 
   private static class GenerateMutations extends DoFn<Long, Mutation> {
